requirements.txt
aiohttp==3.10.10
nest_asyncio==1.6.0
python-socketio==5.11.4
pandas==2.2.3
openpyxl==3.1.5